Improve sidenote layout and rendering
Now long sidenotes don't burden the related paragraph with a break after it.

On the other hand, if such sidenotes are too long, and the next paragraph also
has them, they will overlap destructively.

Luckily this can be tested and avoided, since the sidenotes-to-the-side layout
is fixed-width.

Some miscellaneous typography improvements are lumped in.