Implement multi-site support
This also implements database template storage, along with some changes
to RPC, improvements to the database layer, and other miscellania.

The engine is getting nearer to a point where I'm actually happy with
it, and hopefully that's not too far.

The rearchitecting also changes the database model -- now all app data
resides within a single database. This is closer to the model used by
some other blogging systems, and is actually a simpler model to work
with.

Flexible settings have been scratched in favour of a simpler database
schema which stores all site-related settings in one place. The config
database is changed into a static PHP file which is saved upon changes.
Right now it's used for a single item, namely, the HMAC key, so the
rewriting shouldn't occur at all.

DbObject takes on more responsibilities and is becoming more of a
makeshift ORM.

Also add an Oxford comma within API.txt.